pyproject.toml
[project]
name = "chatbot-de24"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.123.8",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "markdown>=3.10",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.26.0",
    "python-dotenv>=1.2.1",
]
